+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Modstache Change Log
 
+### 2021-06-26 - Version 1.0.4
+
+Removed reactive bindings on array deletions
+Added options to enable/disable reactive binding in stache attributes
+
 ### 2021-03-26 - Version 1.0.3
 
 Prioritized element properties over attributes